extract helpers to prevent code duplication
