+++ EssentialApp/EssentialAppTests/FeedImageLoaderCacheDecoratorTests.swift
@@ -21,11 +21,10 @@ final class FeedImageLoaderCacheDecorater: FeedImageDataLoader {
     }
 }
 
-final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
+final class FeedImageLoaderCacheDecoratorTests: XCTestCase, FeedImageDataLoaderTestable  {
     func test_loadImage_deliversImageOnLoaderSuccess() {
         let feedImageData = anydata()
-        let loaderSpy = ImageLoaderSpy(result: .success(feedImageData))
-        let sut = FeedImageLoaderCacheDecorater(decoratee: loaderSpy)
+        let (sut, loaderSpy) = makeSUT(loaderResult: .success(feedImageData))
         
         expect(sut, toFinishWith: .success(feedImageData)) {
             loaderSpy.completeAtIndex()
@@ -34,8 +33,7 @@ final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
     
     func test_loadImage_deliversErrorOnLoaderFailure() {
         let error = anyNSError()
-        let loaderSpy = ImageLoaderSpy(result: .failure(error))
-        let sut = FeedImageLoaderCacheDecorater(decoratee: loaderSpy)
+        let (sut, loaderSpy) = makeSUT(loaderResult: .failure(error))
         
         expect(sut, toFinishWith: .failure(error)) {
             loaderSpy.completeAtIndex()
@@ -43,66 +41,13 @@ final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
     }
     
     // MARK: - Helpers
-    private func expect(
-        _ sut: FeedImageDataLoader,
-        toFinishWith expectedResult: FeedImageDataLoader.Result,
-        when action: () -> Void,
-        file: StaticString = #filePath,
-        line: UInt = #line
-    ) {
-        let expectation = self.expectation(description: "wait for image data to load")
-        _ = sut.loadImageData(from: anyURL()) { actualResult in
-            switch (actualResult, expectedResult) {
-            case (.success(let actualData), .success(let expectedData)):
-                XCTAssertEqual(actualData, expectedData, file: file, line: line)
-            case (.failure(let actualError as NSError), .failure(let expectedError as NSError)):
-                XCTAssertEqual(actualError, expectedError, file: file, line: line)
-            default:
-                XCTFail("expected \(expectedResult), got \(actualResult)", file: file, line: line)
-            }
-            expectation.fulfill()
-        }
+    private func makeSUT(loaderResult: FeedImageDataLoader.Result, file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, loader: ImageLoaderSpy) {
+        let loader = ImageLoaderSpy(result: loaderResult)
+        let sut = FeedImageLoaderCacheDecorater(decoratee: loader)
         
-        action()
+        trackMemoryLeaks(loader, file: file, line: line)
+        trackMemoryLeaks(sut, file: file, line: line)
         
-        wait(for: [expectation], timeout: 1.0)
-    }
-    
-    private class ImageLoaderSpy: FeedImageDataLoader {
-        private struct Task: FeedImageDataLoaderTask {
-            private var callback: (() -> Void)
-            
-            init(callback: @escaping () -> Void) {
-                self.callback = callback
-            }
-            
-            func cancel() {
-                callback()
-            }
-        }
-        
-        private let result: FeedImageDataLoader.Result
-        private var messages = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-        private(set) var cancelURLs: [URL] = []
-        
-        var loadedImageURLs: [URL] {
-            messages.map { $0.url }
-        }
-        
-        init(result: FeedImageDataLoader.Result) {
-            self.result = result
-        }
-        
-        func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-            messages.append((url, completion))
-            return Task(callback: { [weak self] in
-                self?.cancelURLs.append(url)
-            })
-        }
-        
-        func completeAtIndex(_ index: Int = 0) {
-            messages[index].completion(result)
-        }
-      
+        return (sut, loader)
     }
 }
